+++ components/SideToolbar/SideToolbar.tsx
@@ -11,6 +11,7 @@ import ListItemText from '@material-ui/core/ListItemText'
 import { useStyles } from './uiStyles'
 import { SimulatorContext } from '../../state/context'
 import { addClient, addService, addResource } from '../../state/reducer'
+const { flatten } = require('ramda')
 
 export interface ISideToolbar {
   isOpen: boolean
@@ -25,7 +26,7 @@ export interface Item {
 export const SideToolbar: FC<ISideToolbar> = (props) => {
   const theme = useTheme()
   const classes = useStyles()
-  const { dispatch } = useContext(SimulatorContext)
+  const { dispatch, state } = useContext(SimulatorContext)
 
   const listItems: Item[] = [
     {
@@ -77,6 +78,13 @@ export const SideToolbar: FC<ISideToolbar> = (props) => {
             concurrency: 5,
           })
         )
+      },
+    },
+    {
+      text: 'Play',
+      action: () => {
+        const items = flatten([state.clients, state.resources, state.services])
+        alert(JSON.stringify(items))
       },
     },
   ]
