+++ package.json
@@ -16,6 +16,7 @@
     "add": "^2.0.6",
     "next": "10.2.3",
     "or": "^0.2.0",
+    "ramda": "^0.27.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-xarrows": "^1.7.0",
